refactor: move assembler functions into its own struct
